feat: introduce TagGroup component and replace CategorySection in ActivitiesPage
