+++ src/app/(pages)/[location]/activities/page.tsx
@@ -1,9 +1,11 @@
-import CategorySection from '@/components/CategorySection';
+import TagGroup from '@/components/TagGroup';
+import CategorySection from '@/components/CategorySection/CategorySection';
+import Carousel from '@/components/Carousel';
 import { capitalise } from '@/utils/capitalise';
 import fetchActivities from '@/utils/fetchActivities';
 import { ActivitiesPageParams } from '@/types/pageParams';
-import Carousel from '@/components/Carousel';
 import slidesData from '@/data/carousel.json';
+import CategoryData from '@/data/categories.json';
 
 const ActivitiesPage = async ({ params }: { params: ActivitiesPageParams }) => {
   const { location } = await params;
@@ -17,10 +19,11 @@ const ActivitiesPage = async ({ params }: { params: ActivitiesPageParams }) => {
 
       {/* Main Content */}
       <div className="container mx-auto py-8">
+        <CategorySection categories={CategoryData.categories} />
         {Object.entries(activities).length > 0 ? (
           Object.entries(activities).map(
             ([tagGroup, { activities: groupActivities, tags, description }]) => (
-              <CategorySection
+              <TagGroup
                 key={tagGroup}
                 title={capitalise(tagGroup)}
                 description={description}
