--- src/components/CategorySection/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CategorySection';
