+++ src/data/categories.json
@@ -1,0 +1,52 @@
+{
+  "categories": [
+    {
+      "id": 1,
+      "name": "Sports",
+      "description": "Engage in physical activities, from team sports to individual challenges.",
+      "icon": "MdSportsSoccer"
+    },
+    {
+      "id": 2,
+      "name": "Games",
+      "description": "Enjoy a variety of games, including board games, arcade games, and more.",
+      "icon": "FaGamepad"
+    },
+    {
+      "id": 3,
+      "name": "Attractions",
+      "description": "Visit popular tourist attractions, landmarks, and points of interest.",
+      "icon": "FaLandmark"
+    },
+    {
+      "id": 4,
+      "name": "Outdoor Adventures",
+      "description": "Experience the thrill of outdoor activities like hiking, kayaking, and camping.",
+      "icon": "GiMountainClimbing"
+    },
+    {
+      "id": 5,
+      "name": "Entertainment",
+      "description": "Enjoy concerts, theatre performances, cinemas, and live shows.",
+      "icon": "MdMovie"
+    },
+    {
+      "id": 6,
+      "name": "Food & Drink",
+      "description": "Explore local restaurants, bars, and unique dining experiences.",
+      "icon": "FaUtensils"
+    },
+    {
+      "id": 7,
+      "name": "Wellness & Relaxation",
+      "description": "Unwind with spas, yoga, meditation, and other wellness activities.",
+      "icon": "GiMeditation"
+    },
+    {
+      "id": 8,
+      "name": "Workshops & Classes",
+      "description": "Learn new skills with cooking classes, art workshops, and more.",
+      "icon": "FaChalkboardTeacher"
+    }
+  ]
+}
